Rename the module to flexclient
Rename the module to flexclient, more consistent with python package naming conventions.
Set version back to 0.1
