+++ README.md
@@ -6,7 +6,7 @@ The FLEX-6000 series radios are software defined "radio Servers" with components
 
 The radio is used via a gigabit ethernet port, which allows users to send commands and open further connections to receive and transmit.
 
-Fles Radio has released an open API, but it is restrictive due to the fact it is built with the .NET framework.
+Flex Radio has released an open API, but it is restrictive due to the fact it is built with the .NET framework.
 Building a new API enables cross-platform uses of the SDR, opening many more opportunities to use its potential.
 
 The initial purpose of this module was to support gnuradio blocks which use the radio, but it can also be used to develop clients in pure python.
